[DE190293] feat(online-judge): implement code submission workflow, sc…

+++ backend/src/main/java/com/swp391/coding_platform/entity/problem/ProblemSubmissionDetailEntity.java
@@ -33,7 +33,7 @@ public class ProblemSubmissionDetailEntity {
     String token;
 
     @Column(name = "execution_time")
-    Double executionTime;
+    Integer executionTime;
 
     @Column(name = "memory_used")
     Integer memoryUsed;
